GH-2 Clean up
